pyproject.toml
[tool.poetry]
name = "podly"
version = "0.1.0"
description = "Condensing your favorite podcasts."
authors = ["scott susanto <susantoscott@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
reflex = "^0.4.8.post1"
supabase = "^2.4.3"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
